+++ crates/astria-sequencer/src/authority/storage/keys.rs
@@ -1,6 +1,7 @@
 pub(in crate::authority) const SUDO: &str = "authority/sudo";
 pub(in crate::authority) const VALIDATOR_SET: &str = "authority/validator_set";
 pub(in crate::authority) const VALIDATOR_UPDATES: &str = "authority/validator_updates";
+pub(in crate::authority) const VALIDATOR_NAMES_PREFIX: &str = "authority/validator_names";
 
 #[cfg(test)]
 mod tests {
@@ -13,6 +14,7 @@ mod tests {
         insta::assert_snapshot!("sudo_address_key", SUDO);
         insta::assert_snapshot!("validator_set_key", VALIDATOR_SET);
         insta::assert_snapshot!("validator_updates_key", VALIDATOR_UPDATES);
+        insta::assert_snapshot!("validator_names_prefix", VALIDATOR_NAMES_PREFIX);
     }
 
     #[test]
@@ -20,5 +22,6 @@ mod tests {
         assert!(SUDO.starts_with(COMPONENT_PREFIX));
         assert!(VALIDATOR_SET.starts_with(COMPONENT_PREFIX));
         assert!(VALIDATOR_UPDATES.starts_with(COMPONENT_PREFIX));
+        assert!(VALIDATOR_NAMES_PREFIX.starts_with(COMPONENT_PREFIX));
     }
 }