+++ crates/astria-sequencer/CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add validator update action with name [#1679](https://github.com/astriaorg/astria/pull/1679)
+
 ### Changed
 
 - Index all event attributes [#1786](https://github.com/astriaorg/astria/pull/1786).